# Conflicts:
#	api-impl/src/main/java/com/toutiao/web/apiimpl/impl/Intelligence/IntelligenceFindHouseController.java
#	api-impl/src/main/resources/templates/esf/esf-list.ftl
#	api-impl/src/main/resources/templates/search.ftl
#	core-service/src/main/java/com/toutiao/web/service/intelligence/impl/IntelligenceFindHouseServiceImpl.java
#	dao/src/main/resources/mapper/officeweb/IntelligenceFindhouseMapper.xml

+++ dao/src/main/java/com/toutiao/web/dao/entity/officeweb/TotalRoomRatio.java
@@ -9,7 +9,7 @@ public class TotalRoomRatio {
 
     private Integer room;
 
-    private BigDecimal ratio;
+    private Double ratio;
 
     /**
      * 用户画像类型1,2,3,4,5,6,7
